pyproject.toml
[project]
name = "local-notebooklm-discord"
version = "0.1.0"
description = "Local NotebookLM-like Discord bot with file/web/youtube ingestion and RAG"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "discord.py>=2.3.2",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.4",
  "httpx>=0.27.0",
  "chromadb>=0.5.5",
  "sentence-transformers>=3.0.1",
  "pypdf>=4.2.0",
  "beautifulsoup4>=4.12.3",
  "readability-lxml>=0.8.1",
  "youtube-transcript-api>=0.6.2",
  "python-dotenv>=1.0.1",
  "lxml[html_clean]>=5.2.2"
]

[project.scripts]
local-notebooklm = "local_notebooklm.app:main"

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
